Reorganize file structure

+++ src/ui/window.c
@@ -7,6 +7,8 @@ void C4_Window_GetScaledSize(SDL_Window* window, float uiScale, int* w, int* h) 
         uiScale = 0.001f;
     }
 
-    if (w) *w = SDL_roundf((float)*w / uiScale);
-    if (h) *h = SDL_roundf((float)*h / uiScale);
+    if (w)
+        *w = (int)SDL_roundf((float)*w / uiScale);
+    if (h)
+        *h = (int)SDL_roundf((float)*h / uiScale);
 }
